chore: refactor so APIs can be mocked
